Graphics.Gloss.Data.Picture

Documentation

data Picture

A 2D picture

Constructors

Blank
A blank picture, with nothing in it.

Polygon Path
A convex polygon filled with a solid color.

Line Path
A line along an arbitrary path.

Circle Float
A circle with the given radius.

ThickCircle Float Float
A circle with the given radius and thickness. If the thickness is 0 then this is equivalent to Circle.

Arc Float Float Float
A circular arc drawn counter-clockwise between two angles (in degrees) at the given radius.

ThickArc Float Float Float Float
A circular arc drawn counter-clockwise between two angles (in degrees), with the given radius and thickness. If the thickness is 0 then this is equivalent to Arc.

Text String
Some text to draw with a vector font.

Bitmap BitmapData
A bitmap image.

BitmapSection Rectangle BitmapData
A subsection of a bitmap image where the first argument selects a sub section in the bitmap, and second argument determines the bitmap data.

Color Color Picture
A picture drawn with this color.

Translate Float Float Picture
A picture translated by the given x and y coordinates.

Rotate Float Picture
A picture rotated clockwise by the given angle (in degrees).

Scale Float Float Picture
A picture scaled by the given x and y factors.

Pictures [Picture]
A picture consisting of several others.

type Point = (Float, Float)

A point on the x-y plane.

type Vector = Point

A vector can be treated as a point, and vis-versa.

type Path = [Point]

A path through the x-y plane.

Aliases for Picture constructors

blank :: Picture

A blank picture, with nothing in it.

polygon :: Path -> Picture

A convex polygon filled with a solid color.

line :: Path -> Picture

A line along an arbitrary path.

circle :: Float -> Picture

A circle with the given radius.

thickCircle :: Float -> Float -> Picture

A circle with the given thickness and radius. If the thickness is 0 then this is equivalent to Circle.

arc :: Float -> Float -> Float -> Picture

A circular arc drawn counter-clockwise between two angles (in degrees) at the given radius.

thickArc :: Float -> Float -> Float -> Float -> Picture

A circular arc drawn counter-clockwise between two angles (in degrees), with the given radius and thickness. If the thickness is 0 then this is equivalent to Arc.

text :: String -> Picture

Some text to draw with a vector font.

bitmap :: BitmapData -> Picture

A bitmap image

bitmapSection :: Rectangle -> BitmapData -> Picture

a subsection of a bitmap image first argument selects a sub section in the bitmap second argument determines the bitmap data

color :: Color -> Picture -> Picture

A picture drawn with this color.

translate :: Float -> Float -> Picture -> Picture

A picture translated by the given x and y coordinates.

rotate :: Float -> Picture -> Picture

A picture rotated clockwise by the given angle (in degrees).

scale :: Float -> Float -> Picture -> Picture

A picture scaled by the given x and y factors.

pictures :: [Picture] -> Picture

A picture consisting of several others.

Compound shapes

lineLoop :: Path -> Picture

A closed loop along a path.

circleSolid :: Float -> Picture

A solid circle with the given radius.

arcSolid :: Float -> Float -> Float -> Picture

A solid arc, drawn counter-clockwise between two angles (in degrees) at the given radius.

sectorWire :: Float -> Float -> Float -> Picture

A wireframe sector of a circle. An arc is draw counter-clockwise from the first to the second angle (in degrees) at the given radius. Lines are drawn from the origin to the ends of the arc.

rectanglePath

Arguments

:: Float
width of rectangle

-> Float
height of rectangle

-> Path

A path representing a rectangle centered about the origin

rectangleWire :: Float -> Float -> Picture

A wireframe rectangle centered about the origin.

rectangleSolid :: Float -> Float -> Picture

A solid rectangle centered about the origin.

rectangleUpperPath :: Float -> Float -> Path

A path representing a rectangle in the y > 0 half of the x-y plane.

rectangleUpperWire :: Float -> Float -> Picture

A wireframe rectangle in the y > 0 half of the x-y plane.

rectangleUpperSolid :: Float -> Float -> Picture

A solid rectangle in the y > 0 half of the x-y plane.