Graphics.Gloss.Data.Color

Description

Predefined and custom colors.

Color data type

data Color

An abstract color value. We keep the type abstract so we can be sure that the components are in the required range. To make a custom color use makeColor.

makeColor

Arguments

:: Float
Red component.

-> Float
Green component.

-> Float
Blue component.

-> Float
Alpha component.

-> Color

Make a custom color. All components are clamped to the range [0..1].

makeColorI :: Int -> Int -> Int -> Int -> Color

Make a custom color. All components are clamped to the range [0..255].

rgbaOfColor :: Color -> (Float, Float, Float, Float)

Take the RGBA components of a color.

Color functions

mixColors

Arguments

:: Float
Proportion of first color.

-> Float
Proportion of second color.

-> Color
First color.

-> Color
Second color.

-> Color
Resulting color.

Mix two colors with the given ratios.

addColors :: Color -> Color -> Color

Add RGB components of a color component-wise, then normalise them to the highest resulting one. The alpha components are averaged.

dim :: Color -> Color

Make a dimmer version of a color, scaling towards black.

bright :: Color -> Color

Make a brighter version of a color, scaling towards white.

light :: Color -> Color

Lighten a color, adding white.

dark :: Color -> Color

Darken a color, adding black.

withRed :: Float -> Color -> Color

Set the red value of a Color.

withGreen :: Float -> Color -> Color

Set the green value of a Color.

withBlue :: Float -> Color -> Color

Set the blue value of a Color.

withAlpha :: Float -> Color -> Color

Set the alpha value of a Color.

Pre-defined colors

greyN :: Float -> Color

A greyness of a given order.

Range is 0 = black, to 1 = white.

black :: Color

white :: Color

Primary

red :: Color

green :: Color

blue :: Color

Secondary

yellow :: Color

cyan :: Color

magenta :: Color

Tertiary

rose :: Color

violet :: Color

azure :: Color

aquamarine :: Color

chartreuse :: Color

orange :: Color