Graphics.Gloss.Data.Controller

Documentation

data Controller

Functions to asynchronously control a Gloss display.

Constructors

Controller

Fields

controllerSetRedraw :: IO ()

Indicate that we want the picture to be redrawn.

controllerModifyViewPort :: (ViewPort -> IO ViewPort) -> IO ()

Modify the current viewport, also indicating that it should be redrawn.