Graphics.Gloss.Data.ViewPort

Documentation

data ViewPort

The ViewPort represents the global transformation applied to the displayed picture. When the user pans, zooms, or rotates the display then this changes the ViewPort.

Constructors

ViewPort

Fields

viewPortTranslate :: !(Float, Float)

Global translation.

viewPortRotate :: !Float

Global rotation (in degrees).

viewPortScale :: !Float

Global scaling (of both x and y coordinates).

viewPortInit :: ViewPort

The initial state of the viewport.

applyViewPortToPicture :: ViewPort -> Picture -> Picture

Translates, rotates, and scales an image according to the ViewPort.

invertViewPort :: ViewPort -> Point -> Point

Takes a point using screen coordinates, and uses the ViewPort to convert it to Picture coordinates. This is the inverse of applyViewPortToPicture for points.