Graphics.Gloss.Data.ViewState

Documentation

data Command

The commands suported by the view controller.

Constructors

CRestore

CTranslate

CRotate

CScale

CBumpZoomOut

CBumpZoomIn

CBumpLeft

CBumpRight

CBumpUp

CBumpDown

CBumpClockwise

CBumpCClockwise

type CommandConfig = [(Command, [(Key, Maybe Modifiers)])]

defaultCommandConfig :: CommandConfig

The default commands. Left click pans, wheel zooms, right click rotates, "r" key resets.

data ViewState

State for controlling the viewport. These are used by the viewport control component.

Constructors

ViewState

Fields

viewStateCommands :: !(Map Command [(Key, Maybe Modifiers)])

The command list for the viewport controller. These can be safely overwridden at any time by deleting or adding entries to the list. Entries at the front of the list take precedence.

viewStateScaleStep :: !Float

How much to scale the world by for each step of the mouse wheel.

viewStateRotateFactor :: !Float

How many degrees to rotate the world by for each pixel of x motion.

viewStateScaleFactor :: !Float

Ratio to scale the world by for each pixel of y motion.

viewStateTranslateMark :: !(Maybe (Float, Float))

During viewport translation, where the mouse was clicked on the window to start the translate.

viewStateRotateMark :: !(Maybe (Float, Float))

During viewport rotation, where the mouse was clicked on the window to starte the rotate.

viewStateScaleMark :: !(Maybe (Float, Float))

During viewport scale, where the mouse was clicked on the window to start the scale.

viewStateViewPort :: ViewPort

The current viewport.

viewStateInit :: ViewState

The initial view state.

viewStateInitWithConfig :: CommandConfig -> ViewState

Initial view state, with user defined config.

updateViewStateWithEvent :: Event -> ViewState -> ViewState

Apply an event to a ViewState.

updateViewStateWithEventMaybe :: Event -> ViewState -> Maybe ViewState

Like updateViewStateWithEvent, but returns Nothing if no update was needed.