Graphics.Gloss.Interface.Pure.Animate

Description

Display mode is for drawing a static picture.

Documentation

module Graphics.Gloss.Data.Display

module Graphics.Gloss.Data.Picture

module Graphics.Gloss.Data.Color

animate

Arguments

:: Display
Display mode.

-> Color
Background color.

-> (Float -> Picture)
Function to produce the next frame of animation. It is passed the time in seconds since the program started.

-> IO ()

Open a new window and display the given animation.

Once the window is open you can use the same commands as with display.